% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAPIT.Blink.R
\name{Blink}
\alias{Blink}
\title{Blink}
\usage{
Blink(
  Y = NULL,
  QTN.position = NULL,
  GD = NULL,
  GM = NULL,
  CV = NULL,
  DPP = 1e+08,
  kinship.algorithm = "FARM-CPU",
  file.output = TRUE,
  cutOff = 0.01,
  method.GLM = "FarmCPU.LM",
  Prior = NULL,
  ncpus = 1,
  maxLoop = 10,
  LD = 0.7,
  threshold.output = 1e-04,
  alpha = c(0.01, 0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1),
  WS = c(1, 1000, 10000, 1e+05, 1e+06, 1e+07),
  GP = NULL,
  FDRcut = FALSE,
  maxOut = 10,
  converge = 1,
  iteration.output = FALSE,
  acceleration = 0,
  threshold = NA,
  model = "A",
  MAF.calculate = FALSE,
  plot.style = "FarmCPU",
  p.threshold = NA,
  maf.threshold = 0,
  bound = FALSE,
  method.sub = "reward",
  method.sub.final = "reward",
  stepwise = FALSE,
  BIC.method = "naive",
  LD.wise = FALSE,
  time.cal = FALSE,
  Prediction = FALSE
)
}
\arguments{
\item{Y}{= NULL, data.frame of phenotypic data, column one is sample (taxa), column two is a trait, subsequent columns are other traits.}

\item{QTN.position}{= NULL,}

\item{GD}{= NULL, data.frame of genetic data in 'numerical' format, samples in rows, variants in columns.}

\item{GM}{= NULL, Genetic Map data.frame to provide genomic coordinates for GD}

\item{CV}{= NULL, Covariates}

\item{DPP}{= 100000000,}

\item{kinship.algorithm}{= "FARM-CPU",}

\item{file.output}{= TRUE,}

\item{cutOff}{= 0.01,}

\item{method.GLM}{= "FarmCPU.LM",}

\item{Prior}{= NULL,}

\item{ncpus}{= 1,}

\item{maxLoop}{= 10,}

\item{LD}{= 0.7,}

\item{threshold.output}{= .0001,}

\item{alpha}{= c(.01,.05,.1,.2,.3,.4,.5,.6,.7,.8,.9,1),}

\item{WS}{= c(1e0,1e3,1e4,1e5,1e6,1e7),}

\item{GP}{= NULL,}

\item{FDRcut}{= FALSE,}

\item{maxOut}{= 10,}

\item{converge}{= 1,}

\item{iteration.output}{= FALSE,}

\item{acceleration}{= 0,}

\item{threshold}{= NA,}

\item{model}{= "A",}

\item{MAF.calculate}{= FALSE,}

\item{plot.style}{= "FarmCPU",}

\item{p.threshold}{= NA,}

\item{maf.threshold}{= 0,}

\item{bound}{= FALSE,}

\item{method.sub}{= "reward",}

\item{method.sub.final}{= "reward",}

\item{stepwise}{= FALSE,}

\item{BIC.method}{= "naive",}

\item{LD.wise}{= FALSE,}

\item{time.cal}{= FALSE,}

\item{Prediction}{=  F}
}
\value{
list(GWAS=GWAS,myGLM=myGLM,PEV = PEV,seqQTN=seqQTN)
}
\description{
Blink
}
