% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAPIT.FarmCPU.R
\name{FarmCPU}
\alias{FarmCPU}
\title{FarmCPU}
\usage{
FarmCPU(
  Y = NULL,
  GD = NULL,
  GM = NULL,
  CV = NULL,
  GP = NULL,
  Yt = NULL,
  DPP = 1e+06,
  kinship.algorithm = "FARM-CPU",
  file.output = TRUE,
  cutOff = 0.01,
  method.GLM = "FarmCPU.LM",
  method.sub = "reward",
  method.sub.final = "reward",
  method.bin = "static",
  bin.size = c(5e+05, 5e+06, 5e+07),
  bin.selection = seq(10, 100, 10),
  memo = NULL,
  Prior = NULL,
  ncpus = 1,
  maxLoop = 10,
  threshold.output = 0.01,
  WS = c(1, 1000, 10000, 1e+05, 1e+06, 1e+07),
  alpha = c(0.01, 0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1),
  maxOut = 100,
  QTN.position = NULL,
  converge = 1,
  iteration.output = FALSE,
  acceleration = 0,
  model = "A",
  MAF.calculate = FALSE,
  plot.style = "FarmCPU",
  p.threshold = NA,
  QTN.threshold = 0.01,
  maf.threshold = 0.03,
  ycor = NULL,
  bound = NULL
)
}
\arguments{
\item{Y}{= NULL, a data.frame of phenotype data, first column is sample name, second column is the trait.}

\item{GD}{= NULL,}

\item{GM}{= NULL,}

\item{CV}{= NULL,}

\item{GP}{= NULL,}

\item{Yt}{= NULL,}

\item{DPP}{= 1000000,}

\item{kinship.algorithm}{= "FARM-CPU",}

\item{file.output}{= TRUE,}

\item{cutOff}{= 0.01,}

\item{method.GLM}{= "FarmCPU.LM",}

\item{method.sub}{= "reward",}

\item{method.sub.final}{= "reward",}

\item{method.bin}{= "static",}

\item{bin.size}{= c(5e5,5e6,5e7),}

\item{bin.selection}{= seq(10,100,10),}

\item{memo}{= NULL,}

\item{Prior}{= NULL,}

\item{ncpus}{= 1,}

\item{maxLoop}{= 10,}

\item{threshold.output}{= .01,}

\item{WS}{= c(1e0,1e3,1e4,1e5,1e6,1e7),}

\item{alpha}{= c(.01,.05,.1,.2,.3,.4,.5,.6,.7,.8,.9,1),}

\item{maxOut}{= 100,}

\item{QTN.position}{= NULL,}

\item{converge}{= 1,}

\item{iteration.output}{= FALSE,}

\item{acceleration}{= 0,}

\item{model}{= "A",}

\item{MAF.calculate}{= FALSE,}

\item{plot.style}{= "FarmCPU",}

\item{p.threshold}{= NA,}

\item{QTN.threshold}{= 0.01,}

\item{maf.threshold}{= 0.03,}

\item{ycor}{= NULL,}

\item{bound}{= NULL}
}
\value{
A list.
}
\description{
FarmCPU: GWAS and GS by using FarmCPU method
}
\examples{
\dontrun{
myPhenoFile <- system.file("extdata", "mdp_traits.txt.gz", package = "GAPIT3")
myPhenotypes <- read.table(myPhenoFile, header = TRUE)
myFarmCPU <- FarmCPU(myPhenotypes[, 1:2])
}


}
\author{
Xiaolei Liu and Zhiwu Zhang
}
