% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAPIT.Compression.Visualization.R
\name{GAPIT.Compression.Visualization}
\alias{GAPIT.Compression.Visualization}
\title{GAPIT.Compression.Visualization}
\usage{
GAPIT.Compression.Visualization(
  Compression = Compression,
  name.of.trait = name.of.trait,
  file.output = FALSE
)
}
\arguments{
\item{Compression}{= Compression,}

\item{name.of.trait}{= name.of.trait}

\item{file.output}{= TRUE, should output be automatically written to file.}
}
\value{
An invisible NULL.
}
\description{
Compression visualization
}
\author{
Alex Lipka and Zhiwu Zhang
}
