% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAPIT.ID.R
\name{GAPIT.ID}
\alias{GAPIT.ID}
\title{GAPIT.ID}
\usage{
GAPIT.ID(
  DP = NULL,
  IC = NULL,
  SS = NULL,
  RS = NULL,
  cutOff = 0.01,
  DPP = 1e+05,
  Create.indicator = FALSE,
  FDR.Rate = 1,
  QTN.position = NULL,
  plot.style = "Oceanic",
  file.output = TRUE,
  SNP.MAF = 0,
  CG = NULL,
  plot.bin = 10^9
)
}
\value{
An invisible NULL.
}
\description{
GAPIT.ID
}
