% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAPIT.Main.R
\name{GAPIT.Main}
\alias{GAPIT.Main}
\title{GAPIT.Main}
\usage{
GAPIT.Main(
  Y,
  G = NULL,
  GD = NULL,
  GM = NULL,
  KI = NULL,
  Z = NULL,
  CV = NULL,
  CV.Inheritance = NULL,
  SNP.P3D = TRUE,
  GP = NULL,
  GK = NULL,
  group.from = 1e+06,
  group.to = 1,
  group.by = 10,
  kinship.cluster = "average",
  kinship.group = "Mean",
  kinship.algorithm = NULL,
  DPP = 50000,
  ngrid = 100,
  llin = -10,
  ulim = 10,
  esp = 1e-10,
  GAPIT3.output = TRUE,
  file.path = NULL,
  file.from = NULL,
  file.to = NULL,
  file.total = NULL,
  file.fragment = 512,
  file.G = NULL,
  file.Ext.G = NULL,
  file.GD = NULL,
  file.GM = NULL,
  file.Ext.GD = NULL,
  file.Ext.GM = NULL,
  SNP.MAF = 0,
  FDR.Rate = 1,
  SNP.FDR = 1,
  SNP.effect = "Add",
  SNP.impute = "Middle",
  PCA.total = 0,
  GAPIT.Version = GAPIT.Version,
  name.of.trait,
  GT = NULL,
  SNP.fraction = 1,
  seed = 123,
  BINS = 20,
  SNP.test = TRUE,
  SNP.robust = "FaST",
  LD.chromosome = NULL,
  LD.location = NULL,
  LD.range = NULL,
  model = model,
  bin.from = 10000,
  bin.to = 5e+06,
  bin.by = 1000,
  inclosure.from = 10,
  inclosure.to = 1000,
  inclosure.by = 10,
  SNP.permutation = FALSE,
  SNP.CV = NULL,
  NJtree.group = NJtree.group,
  NJtree.type = NJtree.type,
  plot.bin = plot.bin,
  genoFormat = NULL,
  hasGenotype = NULL,
  byFile = NULL,
  fullGD = NULL,
  PC = NULL,
  GI = NULL,
  Timmer = NULL,
  Memory = NULL,
  sangwich.top = NULL,
  sangwich.bottom = NULL,
  QC = TRUE,
  GTindex = NULL,
  LD = 0.05,
  file.output = TRUE,
  cutOff = 0.05,
  Model.selection = FALSE,
  Create.indicator = FALSE,
  QTN = NULL,
  QTN.round = 1,
  QTN.limit = 0,
  QTN.update = TRUE,
  QTN.method = "Penalty",
  Major.allele.zero = FALSE,
  QTN.position = NULL,
  SUPER_GD = NULL,
  SUPER_GS = SUPER_GS,
  plot.style = "Beach",
  CG = CG,
  chor_taxa = chor_taxa
)
}
\arguments{
\item{Y}{data.frame of phenotype data, samples in rows, traits in columns}

\item{G}{data.frame of genotypic data, HAPMAP format}

\item{GD}{data.frame of genotypic data}

\item{GM}{genetic map for GD}

\item{KI}{param}

\item{Z}{param}

\item{CV}{param}

\item{CV.Inheritance}{param}

\item{SNP.P3D}{param}

\item{GP}{param}

\item{GK}{param}

\item{group.from}{param}

\item{group.to}{param}

\item{group.by}{param}

\item{kinship.cluster}{param}

\item{kinship.group}{param}
}
\value{
A list
}
\description{
GAPIT.Main
}
