% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAPIT.R
\name{GAPIT}
\alias{GAPIT}
\title{GAPIT Genome Association and Prediction Integrated Tools}
\usage{
GAPIT(
  Y = NULL,
  G = NULL,
  GD = NULL,
  GM = NULL,
  KI = NULL,
  Z = NULL,
  CV = NULL,
  Aver.Dis = 1000,
  buspred = FALSE,
  bin.from = 10000,
  bin.to = 10000,
  bin.by = 10000,
  cutOff = 0.05,
  CV.Extragenetic = 0,
  effectunit = 1,
  file.output = TRUE,
  FDRcut = FALSE,
  group.from = 1e+06,
  group.to = 1e+06,
  group.by = 50,
  Geno.View.output = TRUE,
  h2 = NULL,
  inclosure.from = 10,
  inclosure.to = 10,
  inclosure.by = 10,
  Inter.Plot = FALSE,
  Inter.type = c("m", "q"),
  kinship.cluster = "average",
  kinship.group = "Mean",
  kinship.algorithm = "Zhang",
  lmpred = FALSE,
  model = "MLM",
  maxOut = 100,
  memo = NULL,
  Model.selection = FALSE,
  Multi_iter = TRUE,
  Major.allele.zero = FALSE,
  Multiple_analysis = TRUE,
  num_regwas = 10,
  bin.regwas = 1e+05,
  N4 = FALSE,
  NQTN = NULL,
  N.sig = NULL,
  NJtree.group = NULL,
  NJtree.type = c("fan", "unrooted"),
  output.numerical = FALSE,
  output.hapmap = FALSE,
  QC.Y = FALSE,
  QTN.position = NULL,
  QTN.gs = 0,
  QTNDist = "normal",
  r = 0.25,
  Random.model = TRUE,
  sangwich.top = NULL,
  sangwich.bottom = NULL,
  seq.cutoff = NULL,
  seq.num = 50,
  SNP.P3D = TRUE,
  SNP.effect = "Add",
  SNP.impute = "Middle",
  SNP.fraction = 1,
  SNP.test = TRUE,
  SNP.MAF = 0,
  SNP.FDR = 1,
  testY = NULL,
  plot.bin = 10^5,
  PCA.total = 0,
  PCA.col = NULL,
  PCA.3d = FALSE,
  PCA.legend = NULL,
  PCA.View.output = TRUE,
  Phenotype.View = TRUE,
  Predict.type = "GEBV",
  WS = c(1, 1000, 10000, 1e+05, 1e+06, 1e+07),
  WS0 = 10000
)
}
\arguments{
\item{Y}{data.frame of phenotype data where each row is a sample and each column is a trait, the first column is the sample names}

\item{G}{data.frame of genotypic data in HAPMAP format}

\item{GD}{data.frame of genetic data in numerical format, where each row is a sample and each column is a variant.}

\item{GM}{a data.frame of genomic coordinates for the genetic map}

\item{KI}{an $NxN$ matrix of kinship coefficients}

\item{Z}{an $NxN$ (for MLM) or an $NxN`$ (CMLM) matrix of index, which is made with 0 and 1 value to indicate indivdual belong to each group.}

\item{CV}{Covariance matrix}

\item{buspred}{logical, option for prediction after GWAS。}

\item{bin.from}{integer, minimum number of bin(s) to consider in SUPER}

\item{bin.to}{integer, maximum number of bin(s) to consider in SUPER}

\item{bin.by}{integer, increment for evaluating bin size in SUPER}

\item{cutOff}{numeric value, the threshold for filtering significant markers from all. It would be transfor as Bornferrni cutoff in Manhattan plots.}

\item{effectunit}{numeric value, the effect unit of the first choosed marker in the simulation pheotype. default as 1}

\item{file.output}{logical, whether output all result files.}

\item{FDRcut}{logical, filter pseudo QTN based on FDR cut-off in BLINK}

\item{group.from}{integer, minimum number of group(s) to consider in CMLM}

\item{group.to}{integer, maximum number of group(s) to consider in CMLM}

\item{group.by}{integer, increment for evaluating group size in CMLM}

\item{Geno.View.output}{logical whether to output the Genotype analysis including MAF, heterzygosity, LD decay, and other genotype distribution output.}

\item{h2}{numeric value, to set simulation phenotype heritability. It ranged from 0 to 1 means 0% to 100%.}

\item{inclosure.from}{integer, minimum number of pesudo QTNs to consider in SUPER}

\item{inclosure.to}{integer, maximum number of pesudo QTNs to consider in SUPER}

\item{inclosure.by}{integer, increment for evaluating number of pesudo QTNs in SUPER}

\item{Inter.Plot}{logical, whether to output the interactive Manhattan and QQ plots.}

\item{Inter.type}{Interactive plot type for Manhattan and QQ plots."m" indicate manhattan plot and "q" indicate QQ plot.}

\item{kinship.cluster}{algorithm for calculating kinship centroid (options: "average", "complete", "ward", "single", "mcquitty", "median", and "centroid")}

\item{kinship.group}{method for calculating group membership (options: "Mean", "Max", "Min", and "Median")}

\item{kinship.algorithm}{algorithm to calculate the kinship matrix (options: "VanRaden", "EMMA", "Loiselle", and "Zhang")}

\item{lmpred}{logical (vector), option for seletion of linear model prediction or (and) ABLUP.}

\item{model}{model type to run, (options: "MLM", "GLM", "CMLM", "MMLM", "SUPER", "FarmCPU", "gBLUP",  "cBLUP", and "sBLUP"}

\item{maxOut}{numeric, set the number of markers in the power calculation, the top maxOut number of P values markers should be selected.}

\item{memo}{text, from users to remark for output files.}

\item{Model.selection}{logical, whether evaluate optimum number of CV file. If TRUE, all likelyhood values should be evaluated for each CV combination.}

\item{Multi_iter}{logical, whether add more iterations for FarmCPU, BLINK.}

\item{Major.allele.zero}{logical, whether set major allele as 0, and minor allele as 2, if FALSE, they will be set as reverse.}

\item{Multiple_analysis}{logical, whether to output the mulitple mahattan and QQ plots. default as TRUE}

\item{num_regwas}{numeric, the maximum number of selective significant markers into re-GWAS model.}

\item{NQTN}{numeric value, to set simulation number of QTN. It ranged from 1 to the total markers number.}

\item{NJtree.group}{numeric, set the number of clustering groups in the NJtree plot.}

\item{NJtree.type}{type of neighbor joining tree (options: "fan" and "unrooted")}

\item{output.numerical}{logical,whether output numerical genotype file from HapMap file.}

\item{output.hapmap}{logical,whether output numerical HapMap file from numerical genotype file.}

\item{QTN.position}{numeric vector, set where are the QTNs' position. Its maximun values should be equal to total marker number, and its length should be equal to the NQTN.}

\item{QTNDist}{option for distribution of simulated QTN genetic effect in the simulation,(options: "normal" and "geometry")}

\item{Random.model}{logical, whether ran random model to estimate PVE values for significant markers after GWAS.}

\item{sangwich.top}{Model type to run in the first iteration of SUPER, (options: "MLM", "GLM", "CMLM","Fast-LMM")}

\item{sangwich.bottom}{Model type to run in the last iteration of SUPER, (options: "MLM", "GLM", "CMLM","Fast-LMM")}

\item{seq.cutoff}{numeric value, the threshold for filtering significant markers from all. It would be transfor as Bornferrni cutoff in GGS.}

\item{SNP.P3D}{logical, to use P3D or Not for Testing SNPs}

\item{SNP.effect}{genetic model for coding the SNP effect (options: "Add" (additive), "Dom", "Left", and "Right")}

\item{SNP.impute}{SNP imputation method (options: "Middle", "Major", and "Minor")}

\item{SNP.fraction}{numerical input between 0 and 1, fraction of SNPs Sampled to Estimate Kinship and PCs}

\item{SNP.test}{logical, whether to do GWAS or GS.}

\item{SNP.MAF}{numerical input between 0 and 1, minor allele frequency to filter SNPs in GWAS reports}

\item{SNP.FDR}{numerical input between 0 and 1, false discovery rate for filtering SNPs}

\item{testY}{data.frame of phenotype data in testing population, where each row is a sample and each column is a trait, the first column is the sample names.}

\item{PCA.total}{integer, number of principal components to include in Q matrix (can be zero)}

\item{PCA.col}{list for points color in PCA plot. The total length of PCA.col should be equal to the number of individuals in the GD or G file.}

\item{PCA.3d}{logical, whether output 3D PCA plot.}

\item{PCA.View.output}{logical, whether to output the PCA view}

\item{Predict.type}{option to display which type predicted factor again real phenotype in the GAPIT.Association.Prediction pdf file.(options: "GEBV","BLUP" and "BLUE")}

\item{WS}{numeric or numeric vector, the distance between detected markers and real QTN should be recognized as a real power.}

\item{WS0}{numeric, the cutoff threshold for distance between markers to display in GAPIT.Genotype.Distance_R_Chro.pdf file.}

\item{Aver.Dis=1000}{numeric, average display windowsize in LD decay plot,}
}
\value{
A list
including some of the following elements:MLM, GLM, CMLM, MMLM, SUPER, FarmCPU, gBLUP, cBLUP
}
\description{
GWAS and GS procedure using the Multiple models (General Linear Model, Mixed Linear Model, Compression Mixed Linear Model, SUPER, Multiple Loci Mixed linear Model, FarmCPU, and BLINK)
}
\details{
Genome Association and Prediction Integrated Tools
Available models: MLM, GLM, CMLM, MMLM, SUPER, FarmCPU, gBLUP, cBLUP
}
\examples{
\dontrun{

myPhenoFile <- system.file("extdata", "mdp_traits.txt.gz", package = "GAPIT")
myGenoFile <- system.file("extdata", "mdp_genotype_test.hmp.txt.gz", package = "GAPIT")
myPhenotypes <- read.table(myPhenoFile, header = TRUE)
myGenotypes  <- read.table(myGenoFile, header = FALSE)

myGAPIT <- GAPIT(
  Y = myPhenotypes,
  G = myGenotypes,
  PCA.total = 3,
  file.output = FALSE,
  model = "MLM"
)
}


}
\seealso{
GAPIT.DP(), GAPIT.Phenotype.View(), GAPIT.judge(), GAPIT.IC(), GAPIT.SS(), GAPIT.ID().


library(help = "GAPIT")
}
\author{
Zhiwu Zhang and Jiabo Wang
}
