% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{HAPMAP format data}
\alias{HAPMAP format data}
\title{HAPMAP format data}
\description{
HAPMAP format data
}
\details{
Genotypic data in HAPMAP format.

Variants are in rows.
The first 11 columns contain information about each variant.
GAPIT uses columns 1 (variant name; rs), 3 (chromosome; chrom), and 4 (position; pos).
Subsequent columns are samples.
The example dataset consists of 3093 variants scored for 280 samples.
}
\examples{
 
\dontrun{
myGenoFile <- system.file("extdata", "mdp_genotype_test.hmp.txt.gz", package = "GAPIT3")
myGenotypes  <- read.table(myGenoFile, header = FALSE)
myGenotypes[1:4, 1:11]
myGenotypes[1:4, c(1, 12:20)]
}


}
