% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Phenotypic data}
\alias{Phenotypic data}
\title{Phenotypic data}
\description{
Phenotypic data
}
\details{
Samples are in rows.
The first row is a header.
The first column contains the sample names (Taxa).
Subsequent columns are phenotypic traits.
The example dataset consists of 301 samples scored for three traits.
}
\examples{
 
\dontrun{
myPhenoFile <- system.file("extdata", "mdp_traits.txt.gz", package = "GAPIT3")
myPhenotypes <- read.table(myPhenoFile, header = TRUE)
head(myPhenotypes)
}

}
