% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAPIT.ID.R
\name{GAPIT.ID}
\alias{GAPIT.ID}
\title{GAPIT.ID}
\usage{
GAPIT.ID(
  DP = NULL,
  IC = NULL,
  SS = NULL,
  RS = NULL,
  cutOff = 0.01,
  DPP = 1e+05,
  Create.indicator = FALSE,
  FDR.Rate = 1,
  QTN.position = NULL,
  plot.style = "Oceanic",
  file.output = TRUE,
  SNP.MAF = 0,
  CG = NULL,
  plot.bin = 10^9
)
}
\arguments{
\item{DP}{param a list (118 elements?)}

\item{IC}{param a list (9 elements?)}

\item{SS}{param a list (17 elements?)}

\item{RS}{param}

\item{cutOff}{param}

\item{DPP}{param}

\item{Create.indicator}{param}

\item{FDR.Rate}{param}

\item{QTN.position}{param}

\item{plot.style}{param}

\item{file.output}{param}

\item{SNP.MAF}{param}

\item{CG}{param}

\item{plot.bin}{param}
}
\value{
An invisible NULL.
}
\description{
GAPIT.ID
}
\author{
Zhiwu Zhang and Jiabo Wang
}
