% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAPIT.Main.R
\name{GAPIT.Main}
\alias{GAPIT.Main}
\title{GAPIT.Main}
\usage{
GAPIT.Main(
  Y,
  G = NULL,
  GD = NULL,
  GM = NULL,
  KI = NULL,
  Z = NULL,
  CV = NULL,
  CV.Inheritance = NULL,
  SNP.P3D = TRUE,
  GP = NULL,
  GK = NULL,
  group.from = 1e+06,
  group.to = 1,
  group.by = 10,
  kinship.cluster = "average",
  kinship.group = "Mean",
  kinship.algorithm = NULL,
  DPP = 50000,
  ngrid = 100,
  llin = -10,
  ulim = 10,
  esp = 1e-10,
  GAPIT3.output = TRUE,
  file.path = NULL,
  file.from = NULL,
  file.to = NULL,
  file.total = NULL,
  file.fragment = 512,
  file.G = NULL,
  file.Ext.G = NULL,
  file.GD = NULL,
  file.GM = NULL,
  file.Ext.GD = NULL,
  file.Ext.GM = NULL,
  SNP.MAF = 0,
  FDR.Rate = 1,
  SNP.FDR = 1,
  SNP.effect = "Add",
  SNP.impute = "Middle",
  PCA.total = 0,
  GAPIT.Version = GAPIT.Version,
  name.of.trait,
  GT = NULL,
  SNP.fraction = 1,
  seed = 123,
  BINS = 20,
  SNP.test = TRUE,
  SNP.robust = "FaST",
  LD.chromosome = NULL,
  LD.location = NULL,
  LD.range = NULL,
  model = model,
  bin.from = 10000,
  bin.to = 5e+06,
  bin.by = 1000,
  inclosure.from = 10,
  inclosure.to = 1000,
  inclosure.by = 10,
  SNP.permutation = FALSE,
  SNP.CV = NULL,
  NJtree.group = NJtree.group,
  NJtree.type = NJtree.type,
  plot.bin = plot.bin,
  genoFormat = NULL,
  hasGenotype = NULL,
  byFile = NULL,
  fullGD = NULL,
  PC = NULL,
  GI = NULL,
  Timmer = NULL,
  Memory = NULL,
  sangwich.top = NULL,
  sangwich.bottom = NULL,
  QC = TRUE,
  GTindex = NULL,
  LD = 0.05,
  file.output = TRUE,
  cutOff = 0.05,
  Model.selection = FALSE,
  Create.indicator = FALSE,
  QTN = NULL,
  QTN.round = 1,
  QTN.limit = 0,
  QTN.update = TRUE,
  QTN.method = "Penalty",
  Major.allele.zero = FALSE,
  QTN.position = NULL,
  SUPER_GD = NULL,
  SUPER_GS = SUPER_GS,
  plot.style = "Beach",
  CG = CG,
  chor_taxa = chor_taxa
)
}
\arguments{
\item{Y}{data.frame of phenotype data, samples in rows, traits in columns}

\item{G}{data.frame of genotypic data, HAPMAP format}

\item{GD}{data.frame of genotypic data}

\item{GM}{genetic map for GD}

\item{KI}{param}

\item{Z}{param}

\item{CV}{param}

\item{CV.Inheritance}{param}

\item{SNP.P3D}{param}

\item{GP}{param}

\item{GK}{param}

\item{group.from}{param}

\item{group.to}{param}

\item{group.by}{param}

\item{kinship.cluster}{param}

\item{kinship.group}{param}

\item{kinship.algorithm}{param}

\item{DPP}{param}

\item{ngrid}{param}

\item{llin}{param}

\item{ulim}{param}

\item{esp}{param}

\item{GAPIT3.output}{param}

\item{file.path}{param}

\item{file.from}{param}

\item{file.to}{param}

\item{file.total}{param}

\item{file.fragment}{param}

\item{file.G}{param}

\item{file.Ext.G}{param}

\item{file.GD}{param}

\item{file.GM}{param}

\item{file.Ext.GD}{param}

\item{file.Ext.GM}{param}

\item{SNP.MAF}{param}

\item{FDR.Rate}{param}

\item{SNP.FDR}{param}

\item{SNP.effect}{param}

\item{SNP.impute}{param}

\item{PCA.total}{param}

\item{GAPIT.Version}{param}

\item{name.of.trait}{param}

\item{GT}{param}

\item{SNP.fraction}{param}

\item{seed}{param}

\item{BINS}{param}

\item{SNP.test}{param}

\item{SNP.robust}{param}

\item{LD.chromosome}{param}

\item{LD.location}{param}

\item{LD.range}{param}

\item{model}{param}

\item{bin.from}{param}

\item{bin.to}{param}

\item{bin.by}{param}

\item{inclosure.from}{param}

\item{inclosure.to}{param}

\item{inclosure.by}{param}

\item{SNP.permutation}{param}

\item{SNP.CV}{param}

\item{NJtree.group}{param}

\item{NJtree.type}{param}

\item{plot.bin}{param}

\item{genoFormat}{param}

\item{hasGenotype}{param}

\item{byFile}{param}

\item{fullGD}{param}

\item{PC}{param}

\item{GI}{param}

\item{Timmer}{param}

\item{Memory}{param}

\item{sangwich.top}{param}

\item{sangwich.bottom}{param}

\item{QC}{param}

\item{GTindex}{param}

\item{LD}{param}

\item{file.output}{param}

\item{cutOff}{param}

\item{Model.selection}{param}

\item{Create.indicator}{param}

\item{QTN}{param}

\item{QTN.round}{param}

\item{QTN.limit}{param}

\item{QTN.update}{param}

\item{QTN.method}{param}

\item{Major.allele.zero}{param}

\item{QTN.position}{param}

\item{SUPER_GD}{param}

\item{SUPER_GS}{param}

\item{plot.style}{param}

\item{CG}{param}

\item{chor_taxa}{param}
}
\value{
A list
}
\description{
GAPIT.Main
}
\author{
Zhiwu Zhang and Jiabo Wang
}
