% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAPIT.R
\name{GAPIT}
\alias{GAPIT}
\title{GAPIT Genome Association and Prediction Integrated Tools}
\usage{
GAPIT(
  Y = NULL,
  G = NULL,
  GD = NULL,
  GM = NULL,
  KI = NULL,
  Z = NULL,
  CV = NULL,
  CV.Inheritance = NULL,
  GP = NULL,
  GK = NULL,
  testY = NULL,
  group.from = 1e+06,
  group.to = 1e+06,
  group.by = 20,
  DPP = 1e+05,
  kinship.cluster = "average",
  kinship.group = "Mean",
  kinship.algorithm = "VanRaden",
  buspred = FALSE,
  lmpred = FALSE,
  FDRcut = FALSE,
  bin.from = 10000,
  bin.to = 10000,
  bin.by = 10000,
  inclosure.from = 10,
  inclosure.to = 10,
  inclosure.by = 10,
  SNP.P3D = TRUE,
  SNP.effect = "Add",
  SNP.impute = "Middle",
  PCA.total = 0,
  SNP.fraction = 1,
  seed = NULL,
  BINS = 20,
  SNP.test = TRUE,
  SNP.MAF = 0,
  FDR.Rate = 1,
  SNP.FDR = 1,
  SNP.permutation = FALSE,
  SNP.CV = NULL,
  SNP.robust = "GLM",
  file.from = 1,
  file.to = 1,
  file.total = NULL,
  file.fragment = 99999,
  file.path = NULL,
  file.G = NULL,
  file.Ext.G = NULL,
  file.GD = NULL,
  file.GM = NULL,
  file.Ext.GD = NULL,
  file.Ext.GM = NULL,
  ngrid = 100,
  llim = -10,
  ulim = 10,
  esp = 1e-10,
  LD.chromosome = NULL,
  LD.location = NULL,
  LD.range = NULL,
  PCA.col = NULL,
  PCA.3d = FALSE,
  NJtree.group = NULL,
  NJtree.type = c("fan", "unrooted"),
  sangwich.top = NULL,
  sangwich.bottom = NULL,
  QC = TRUE,
  GTindex = NULL,
  LD = 0.1,
  plot.bin = 10^5,
  file.output = TRUE,
  cutOff = 0.05,
  Model.selection = FALSE,
  output.numerical = FALSE,
  output.hapmap = FALSE,
  Create.indicator = FALSE,
  Multi_iter = FALSE,
  num_regwas = 10,
  opt = "extBIC",
  QTN = NULL,
  QTN.round = 1,
  QTN.limit = 0,
  QTN.update = TRUE,
  QTN.method = "Penalty",
  Major.allele.zero = FALSE,
  Random.model = FALSE,
  method.GLM = "FarmCPU.LM",
  method.sub = "reward",
  method.sub.final = "reward",
  method.bin = "static",
  bin.size = c(1e+06),
  bin.selection = c(10, 20, 50, 100, 200, 500, 1000),
  memo = NULL,
  Prior = NULL,
  ncpus = 1,
  maxLoop = 3,
  threshold.output = 0.01,
  Inter.Plot = FALSE,
  Inter.type = c("m", "q"),
  WS = c(1, 1000, 10000, 1e+05, 1e+06, 1e+07),
  alpha = c(0.01, 0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1),
  maxOut = 100,
  QTN.position = NULL,
  CG = NULL,
  converge = 1,
  iteration.output = FALSE,
  acceleration = 0,
  iteration.method = "accum",
  PCA.View.output = TRUE,
  Geno.View.output = TRUE,
  plot.style = "Oceanic",
  SUPER_GD = NULL,
  SUPER_GS = FALSE,
  h2 = NULL,
  NQTN = NULL,
  QTNDist = "normal",
  effectunit = 1,
  category = 1,
  r = 0.25,
  cveff = NULL,
  a2 = 0,
  adim = 2,
  Multiple_analysis = FALSE,
  model = "MLM",
  Para = NULL
)
}
\description{
GAPIT analyzes phenotypic and genotypics data to infer association.
}
\examples{
\dontrun{

myPhenoFile <- system.file("extdata", "mdp_traits.txt.gz", package = "GAPIT3")
myGenoFile <- system.file("extdata", "mdp_genotype_test.hmp.txt.gz", package = "GAPIT3")
myPhenotypes <- read.table(myPhenoFile, header = TRUE)
myGenotypes  <- read.table(myGenoFile, header = FALSE)

myGAPIT <- GAPIT(
  Y = myPhenotypes,
  G = myGenotypes,
  PCA.total = 3,
  model = "MLM"
)
}



}
